Add duplicte column validator.
Checks to see if any columns (other than the id column) are duplicated,
either in one file or across files.

The behind-the-scenes change that *could* have reprecussions is that
this changes how we're reading the CSV files into dataframes. pandas
mangles duplicated column names when reading CSV files; however, we can
get around this by having pandas not interpret the header row and
instead, read it as a normal row and then use it as column headers.

Yes, this is silly. See https://github.com/pandas-dev/pandas/issues/19383