Fix line endings (\r\n) for python 3
Yes, we really do always want Windows line endings for all output CSV
files -- using the files on Windows is more common than on linux and
Windows clients are happier with this.

The behavior we ABSOLUTLEY DO NOT WANT is platform-consistent behavior,
as we potentially have both Windows and linux systems running this, and
we don't want a ton of files with different line endings.