Drop rows with duplicate index values when loading
pandas is totally unwilling to concat() if there are duplicates in the
source dataframe indexes. That's good! Duplicate index values mean you
don't know what data is what.

In this update, we print a warning and drop any rows that have duplicate
index values when making the full masterfile.

Probably a good todo would be to refuse to run `masterfile join` with
any duplicate indexes unless we used a `--force` option or the like.